pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = "Service for managing projects and project related resources."
authors = ["Indoc Research"]

[tool.poetry.dependencies]
python = "^3.9"
common = "0.0.19"
asyncpg = "0.25.0"
fastapi = "0.75.0"
greenlet = "1.1.2"
opentelemetry-api = "1.10.0"
opentelemetry-exporter-jaeger = "1.10.0"
opentelemetry-instrumentation = "0.29b0"
opentelemetry-instrumentation-fastapi = "0.29b0"
psycopg2 = "2.9.3"
pydantic = "1.8.2"
sqlalchemy = "1.4.32"
uvicorn = "0.17.6"

[tool.poetry.dev-dependencies]
alembic = "1.7.7"
faker = "13.3.2"
httpx = "0.21.3"
pytest = "7.1.1"
pytest-asyncio = "0.18.1"
pytest-cov = "3.0.0"
pytest-mock = "3.7.0"
testcontainers = "3.4.2"

[tool.pytest.ini_options]
testpaths = "tests"
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pilot"
url = "https://git.indocresearch.org/api/v4/groups/pilot/-/packages/pypi/simple"
default = true
